Agent Dhara Backend/requirements.txt
# Core
pandas>=1.5.0
pyyaml>=6.0
openpyxl>=3.0.0

# Azure Blob Storage
azure-storage-blob>=12.0.0

# Optional: Azure Identity (managed identity)
azure-identity>=1.12.0

# Azure OpenAI client support
openai>=1.0.0

# Optional HTTP APIs
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
python-multipart>=0.0.6
pydantic>=2.0.0

# MCP — stdio servers:
#   python -m mcp_server.agent_dhara_data_mcp   (all tools)
#   python -m mcp_server.mcp_azure_sql|mcp_azure_blob|mcp_local_fs|mcp_stream
mcp>=1.2.0

# LangGraph — multi-agent orchestration
langgraph>=0.2.0

# Optional: Azure SQL via pythonnet (install pythonnet + .NET SDK separately on your OS)
pythonnet>=3.0.0

# Optional: Word-to-number conversion (e.g. "one hundred" -> 100)
word2number>=1.1
